add mlir lit test command
